Variable fsc(213) is added in commit 13b7c634.
There are two variables with the same name and type in the body of the method `(public int process(List<File> aFiles)` in commit 13b7c634. Both of them are declared as a variable of an enhanced-for. Variable `fsc:FileSetCheck(242)` in the parent commit should be matched with variable `fsc:FileSetCheck(278)` because both have an statement in common in their scope which is `fsc.destroy()`. So, the child commit variable `fsc:FileSetCheck(242)` is a newly added variable.